+++ packages/terra-form-select/src/shared/_Option.jsx
@@ -131,8 +131,7 @@ const Option = ({
       {...customProps}
       disabled={disabled}
       className={optionClassNames}
-      aria-selected={isSelected} // Needed to allow VoiceOver on iOS to announce selected state
-      aria-checked={isSelected} // Needed to allow JAWS to announce "selected" state
+      aria-selected={isSelected} // Needed to allow screen reader to announce selected state
       aria-disabled={disabled}
       tabIndex="0" // eslint-disable-line jsx-a11y/no-noninteractive-tabindex
       data-terra-select-option