+++ packages/terra-form-select/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Added
+  * Added documentation updates for `Single Select`.
+
+* Changed
+  * Removed aria-checked to make screen readers announce Selected in `terra-form-select`.
+  * Updated focus to shift back on select upon key press or mouse click event while using Safari browser in `terra-form-select`.
+
 ## 6.41.0 - (May 9, 2023)
 
 * Added